Late commit!
So far, we have a functional game which is connected to Google
Firebase. Each new username creates a new object in Firebase which
updates per new high score on that specific object. Currently working
on creating a leaderboard that displays users with the top scores,
better layout, and any bugs I come across