Fixed leaderboard display and added loading gif
Leaderboard now shows on page load, without the player having to input
a name first. The leaderboard also will update if player gets a high
score without needing a page refresh. Need to add a cool color
highlight when this happens so it’s more fun! Also added a loading gif
for when the leaderboard hasn’t yet been populated.